Configure for Android
